Split the AMPLRepnVisitor from the NLv2 writer module
