+++ pyomo/contrib/incidence_analysis/config.py
@@ -14,7 +14,7 @@
 import enum
 from pyomo.common.config import ConfigDict, ConfigValue, InEnum
 from pyomo.common.modeling import NOTSET
-from pyomo.repn.plugins.nl_writer import AMPLRepnVisitor
+from pyomo.repn.ampl import AMPLRepnVisitor
 from pyomo.repn.util import FileDeterminism, FileDeterminism_to_SortComponents
 
 
@@ -33,7 +33,7 @@ class IncidenceMethod(enum.Enum):
     """
 
     ampl_repn = 3
-    """Use ``pyomo.repn.plugins.nl_writer.AMPLRepnVisitor``"""
+    """Use ``pyomo.repn.ampl.AMPLRepnVisitor``"""
 
 
 class IncidenceOrder(enum.Enum):
