+++ pyomo/repn/tests/ampl/test_nlv2.py
@@ -53,10 +53,6 @@ _invalid_1j = r'InvalidNumber\((\([-+0-9.e]+\+)?1j\)?\)'
 
 class INFO(object):
     def __init__(self, symbolic=False):
-        if symbolic:
-            self.template = nl_writer.text_nl_debug_template
-        else:
-            self.template = nl_writer.text_nl_template
         self.subexpression_cache = {}
         self.external_functions = {}
         self.var_map = {}
@@ -72,6 +68,7 @@ class INFO(object):
             True,
             None,
         )
+        self.template = self.visitor.template
 
     def __enter__(self):
         return self
